Fixed a bug when closing out all positions.

+++ src/main/java/org/zhuang/trading/api/IBActionsImpl.java
@@ -114,7 +114,7 @@ public class IBActionsImpl implements IBActions {
 
         Contract contract = Contracts.simpleFuture(symbol, contractMonth, exchange);
 
-        Order order = Orders.marketOrder(action, (double)quantity);
+        Order order = Orders.marketOrder(action, quantity);
         order.orderId(orderId);
         order.transmit(true);
